refactor: move tests into a separate folder
