+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
-    "test": "node src/test.js"
+    "test": "node src/app/model/tests/test.js"
   },
   "keywords": [
     "tetris",
