+++ src/app/model/tests/board.move-right.test.js
@@ -1,31 +1,31 @@
 import assert from 'assert/strict'
 
-import { Board } from './board.js'
-import { Shape } from './shape.js'
+import { Board } from '../board.js'
+import { Shape } from '../shape.js'
 
 console.time('Test time')
 
-/* Can Shape move left */
+/* Can Shape move right */
 
 const board = new Board(6, 10)
 
-// left border collisions
+// right border collisions
 
 let typeIndex = 3
-let rotation = 0
+let rotation = 2
 let shape = new Shape(typeIndex, rotation)
-shape.position = { top: 1, left: 3 }
+shape.position = { top: 2, left: 3 }
 let iterations = 0
-while (board.canMoveLeft(shape)) {
+while (board.canMoveRight(shape)) {
   iterations += 1
-  shape.moveLeft()
+  shape.moveRight()
 }
 board.merge(shape)
 assert.deepStrictEqual(board.board, [
   [0, 0, 0, 0, 0, 0],
-  [3, 3, 0, 0, 0, 0],
-  [3, 3, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
+  [0, 0, 0, 0, 3, 3],
+  [0, 0, 0, 0, 3, 3],
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
@@ -33,7 +33,7 @@ assert.deepStrictEqual(board.board, [
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
 ])
-assert.strictEqual(iterations, 3)
+assert.strictEqual(iterations, 1)
 
 board.reset()
 typeIndex = 5
@@ -41,15 +41,15 @@ rotation = 0
 shape = new Shape(typeIndex, rotation)
 shape.position = { top: 2, left: 4 }
 iterations = 0
-while (board.canMoveLeft(shape)) {
+while (board.canMoveRight(shape)) {
   iterations += 1
-  shape.moveLeft()
+  shape.moveRight()
 }
 board.merge(shape)
 assert.deepStrictEqual(board.board, [
   [0, 0, 0, 0, 0, 0],
-  [0, 4, 4, 0, 0, 0],
-  [4, 4, 0, 0, 0, 0],
+  [0, 0, 0, 0, 4, 4],
+  [0, 0, 0, 4, 4, 0],
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
@@ -58,7 +58,7 @@ assert.deepStrictEqual(board.board, [
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
 ])
-assert.strictEqual(iterations, 3)
+assert.strictEqual(iterations, 0)
 
 board.reset()
 typeIndex = 0
@@ -66,15 +66,15 @@ rotation = 2
 shape = new Shape(typeIndex, rotation)
 shape.position = { top: 2, left: 3 }
 iterations = 0
-while (board.canMoveLeft(shape)) {
+while (board.canMoveRight(shape)) {
   iterations += 1
-  shape.moveLeft()
+  shape.moveRight()
 }
 board.merge(shape)
 assert.deepStrictEqual(board.board, [
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
-  [1, 1, 1, 1, 0, 0],
+  [0, 0, 1, 1, 1, 1],
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
@@ -89,18 +89,18 @@ board.reset()
 typeIndex = 0
 rotation = 3
 shape = new Shape(typeIndex, rotation)
-shape.position = { top: 2, left: 5 }
+shape.position = { top: 2, left: 1 }
 iterations = 0
-while (board.canMoveLeft(shape)) {
+while (board.canMoveRight(shape)) {
   iterations += 1
-  shape.moveLeft()
+  shape.moveRight()
 }
 board.merge(shape)
 assert.deepStrictEqual(board.board, [
-  [1, 0, 0, 0, 0, 0],
-  [1, 0, 0, 0, 0, 0],
-  [1, 0, 0, 0, 0, 0],
-  [1, 0, 0, 0, 0, 0],
+  [0, 0, 0, 0, 0, 1],
+  [0, 0, 0, 0, 0, 1],
+  [0, 0, 0, 0, 0, 1],
+  [0, 0, 0, 0, 0, 1],
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
@@ -108,7 +108,7 @@ assert.deepStrictEqual(board.board, [
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
 ])
-assert.strictEqual(iterations, 5)
+assert.strictEqual(iterations, 4)
 
 // t-shape border collisions
 
@@ -118,15 +118,15 @@ rotation = 0
 shape = new Shape(typeIndex, rotation)
 shape.position = { top: 2, left: 3 }
 iterations = 0
-while (board.canMoveLeft(shape)) {
+while (board.canMoveRight(shape)) {
   iterations += 1
-  shape.moveLeft()
+  shape.moveRight()
 }
 board.merge(shape)
 assert.deepStrictEqual(board.board, [
   [0, 0, 0, 0, 0, 0],
-  [0, 5, 0, 0, 0, 0],
-  [5, 5, 5, 0, 0, 0],
+  [0, 0, 0, 0, 5, 0],
+  [0, 0, 0, 5, 5, 5],
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
@@ -135,7 +135,7 @@ assert.deepStrictEqual(board.board, [
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
 ])
-assert.strictEqual(iterations, 2)
+assert.strictEqual(iterations, 1)
 
 board.reset()
 typeIndex = 6
@@ -143,16 +143,16 @@ rotation = 1
 shape = new Shape(typeIndex, rotation)
 shape.position = { top: 2, left: 3 }
 iterations = 0
-while (board.canMoveLeft(shape)) {
+while (board.canMoveRight(shape)) {
   iterations += 1
-  shape.moveLeft()
+  shape.moveRight()
 }
 board.merge(shape)
 assert.deepStrictEqual(board.board, [
   [0, 0, 0, 0, 0, 0],
-  [5, 0, 0, 0, 0, 0],
-  [5, 5, 0, 0, 0, 0],
-  [5, 0, 0, 0, 0, 0],
+  [0, 0, 0, 0, 5, 0],
+  [0, 0, 0, 0, 5, 5],
+  [0, 0, 0, 0, 5, 0],
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
@@ -160,32 +160,7 @@ assert.deepStrictEqual(board.board, [
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
 ])
-assert.strictEqual(iterations, 3)
-
-board.reset()
-typeIndex = 6
-rotation = 2
-shape = new Shape(typeIndex, rotation)
-shape.position = { top: 2, left: 3 }
-iterations = 0
-while (board.canMoveLeft(shape)) {
-  iterations += 1
-  shape.moveLeft()
-}
-board.merge(shape)
-assert.deepStrictEqual(board.board, [
-  [0, 0, 0, 0, 0, 0],
-  [0, 0, 0, 0, 0, 0],
-  [5, 5, 5, 0, 0, 0],
-  [0, 5, 0, 0, 0, 0],
-  [0, 0, 0, 0, 0, 0],
-  [0, 0, 0, 0, 0, 0],
-  [0, 0, 0, 0, 0, 0],
-  [0, 0, 0, 0, 0, 0],
-  [0, 0, 0, 0, 0, 0],
-  [0, 0, 0, 0, 0, 0],
-])
-assert.strictEqual(iterations, 2)
+assert.strictEqual(iterations, 1)
 
 board.reset()
 typeIndex = 6
@@ -193,16 +168,16 @@ rotation = 3
 shape = new Shape(typeIndex, rotation)
 shape.position = { top: 2, left: 3 }
 iterations = 0
-while (board.canMoveLeft(shape)) {
+while (board.canMoveRight(shape)) {
   iterations += 1
-  shape.moveLeft()
+  shape.moveRight()
 }
 board.merge(shape)
 assert.deepStrictEqual(board.board, [
   [0, 0, 0, 0, 0, 0],
-  [0, 5, 0, 0, 0, 0],
-  [5, 5, 0, 0, 0, 0],
-  [0, 5, 0, 0, 0, 0],
+  [0, 0, 0, 0, 0, 5],
+  [0, 0, 0, 0, 5, 5],
+  [0, 0, 0, 0, 0, 5],
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
@@ -215,46 +190,46 @@ assert.strictEqual(iterations, 2)
 // t-shape debris collision
 
 board.board = [
-  [1, 0, 0, 0, 0, 0],
-  [1, 0, 0, 0, 0, 0],
-  [1, 0, 0, 0, 0, 0],
-  [1, 0, 0, 0, 0, 0],
-  [1, 0, 0, 0, 0, 0],
-  [1, 0, 0, 0, 0, 0],
-  [1, 0, 0, 0, 0, 0],
-  [1, 0, 0, 0, 0, 0],
-  [1, 0, 0, 0, 0, 0],
-  [1, 0, 0, 0, 0, 0],
+  [1, 1, 1, 1, 1, 1],
+  [0, 0, 0, 0, 0, 1],
+  [0, 0, 0, 0, 0, 1],
+  [0, 0, 0, 0, 0, 1],
+  [0, 0, 0, 0, 0, 1],
+  [0, 0, 0, 0, 0, 1],
+  [0, 0, 0, 0, 0, 1],
+  [0, 0, 0, 0, 0, 1],
+  [0, 0, 0, 0, 0, 1],
+  [0, 0, 0, 0, 0, 1],
 ]
 typeIndex = 6
 rotation = 3
 shape = new Shape(typeIndex, rotation)
 shape.position = { top: 2, left: 3 }
 iterations = 0
-while (board.canMoveLeft(shape)) {
+while (board.canMoveRight(shape)) {
   iterations += 1
-  shape.moveLeft()
+  shape.moveRight()
 }
 board.merge(shape)
 assert.deepStrictEqual(board.board, [
-  [1, 0, 0, 0, 0, 0],
-  [1, 0, 5, 0, 0, 0],
-  [1, 5, 5, 0, 0, 0],
-  [1, 0, 5, 0, 0, 0],
-  [1, 0, 0, 0, 0, 0],
-  [1, 0, 0, 0, 0, 0],
-  [1, 0, 0, 0, 0, 0],
-  [1, 0, 0, 0, 0, 0],
-  [1, 0, 0, 0, 0, 0],
-  [1, 0, 0, 0, 0, 0],
+  [1, 1, 1, 1, 1, 1],
+  [0, 0, 0, 0, 5, 1],
+  [0, 0, 0, 5, 5, 1],
+  [0, 0, 0, 0, 5, 1],
+  [0, 0, 0, 0, 0, 1],
+  [0, 0, 0, 0, 0, 1],
+  [0, 0, 0, 0, 0, 1],
+  [0, 0, 0, 0, 0, 1],
+  [0, 0, 0, 0, 0, 1],
+  [0, 0, 0, 0, 0, 1],
 ])
 assert.strictEqual(iterations, 1)
 
 board.board = [
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
+  [0, 0, 0, 0, 0, 7],
   [0, 0, 0, 0, 0, 0],
-  [0, 0, 7, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
@@ -267,16 +242,16 @@ rotation = 2
 shape = new Shape(typeIndex, rotation)
 shape.position = { top: 2, left: 3 }
 iterations = 0
-while (board.canMoveLeft(shape)) {
+while (board.canMoveRight(shape)) {
   iterations += 1
-  shape.moveLeft()
+  shape.moveRight()
 }
 board.merge(shape)
 assert.deepStrictEqual(board.board, [
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
-  [0, 0, 5, 5, 5, 0],
-  [0, 0, 7, 5, 0, 0],
+  [0, 0, 5, 5, 5, 7],
+  [0, 0, 0, 5, 0, 0],
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
@@ -289,8 +264,8 @@ assert.strictEqual(iterations, 0)
 board.board = [
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
-  [0, 0, 0, 0, 0, 0],
-  [7, 7, 0, 0, 0, 0],
+  [0, 0, 0, 0, 0, 7],
+  [0, 0, 0, 0, 0, 7],
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
@@ -298,21 +273,21 @@ board.board = [
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
 ]
-typeIndex = 6
+typeIndex = 1
 rotation = 1
 shape = new Shape(typeIndex, rotation)
-shape.position = { top: 2, left: 4 }
+shape.position = { top: 2, left: 2 }
 iterations = 0
-while (board.canMoveLeft(shape)) {
+while (board.canMoveRight(shape)) {
   iterations += 1
-  shape.moveLeft()
+  shape.moveRight()
 }
 board.merge(shape)
 assert.deepStrictEqual(board.board, [
   [0, 0, 0, 0, 0, 0],
-  [0, 0, 5, 0, 0, 0],
-  [0, 0, 5, 5, 0, 0],
-  [7, 7, 5, 0, 0, 0],
+  [0, 0, 0, 0, 2, 2],
+  [0, 0, 0, 0, 2, 7],
+  [0, 0, 0, 0, 2, 7],
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
   [0, 0, 0, 0, 0, 0],
@@ -324,5 +299,5 @@ assert.strictEqual(iterations, 2)
 
 /* Test result */
 
-console.log(`Board move-left tests - \x1b[1;36mDONE\x1b[0m`)
+console.log(`Board move-right tests - \x1b[1;36mDONE\x1b[0m`)
 console.timeEnd('Test time')
