+++ src/app/model/tests/test.js
@@ -1,0 +1,8 @@
+import './utils.test.js'
+import './shape.test.js'
+import './board.test.js'
+import './board.move-down.test.js'
+import './board.move-left.test.js'
+import './board.move-right.test.js'
+import './board.rotate.test.js'
+import './model.test.js'
