--- src/test.js
@@ -1,8 +1,0 @@
-import './app/model/utils.test.js'
-import './app/model/shape.test.js'
-import './app/model/board.test.js'
-import './app/model/board.move-down.test.js'
-import './app/model/board.move-left.test.js'
-import './app/model/board.move-right.test.js'
-import './app/model/board.rotate.test.js'
-import './app/model/model.test.js'
